Moved Battery Monitoring code to Timer Interrupt
Realized that this has to always happen, whether charging, discharging, or just idling. So why not simply increase and decrease the frequency based on mode and put it in a timer interrupt. But then the question becomes "is it fast enough of a sub routine?" I think it is. The routine only checks for a singular message. If it hasn;t received any, it sends out the MBB message. Perhaps we need to use request ID's, but I haven't found a use for this yet. Also need to verify that batteyr modules send back data before message is called again and that messages do not overfill ring buffer.